requirements.txt
asgiref==3.5.2
azure-core==1.26.0
azure-storage-blob==12.13.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
django-storages==1.13.1
heroku==0.1.4
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.1
Pillow==9.2.0
psycopg2==2.9.4
pycparser==2.21
python-dateutil==1.5
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.12
whitenoise==6.2.0
